osa 4 loppuun

+++ osa4/blogilista/src/controllers/blogs.js
@@ -1,11 +1,15 @@
 const blogsRouter = require('express').Router()
+const config = require('../utils/config')
 const Blog = require('../models/blog')
+const User = require('../models/user')
 
-blogsRouter.get('/', (request, response) => {
-    Blog.find({}).then(blogs => {
-      response.json(blogs)
-    })
-  })
+const jwt = require('jsonwebtoken')
+
+blogsRouter.get('/', async (request, response) => {
+  const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 })
+
+    response.json(blogs)
+})
   
   blogsRouter.get('/:id', async (request, response) => {
     const blog = await Blog.findById(request.params.id)
@@ -15,22 +19,37 @@ blogsRouter.get('/', (request, response) => {
           response.status(201).end()
         }
   })
-  
+
   blogsRouter.post('/', async (request, response) => {
     const body = request.body
+    const decodedToken = jwt.verify(request.token, config.SECRET);
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'invalid token'})
+    }
+
+    const user = await User.findById(decodedToken.id)
   
     const blog = new Blog({
         title: body.title,
         author: body.author,
         url: body.url,
         likes: body.likes,
+        user: user._id
     })
   
    const savedBlog = await blog.save()
+        user.blogs = user.blogs.concat(savedBlog._id)
+        await user.save()
+
         response.status(201).json(savedBlog)
   })
   
   blogsRouter.delete('/:id', async (request, response) => {
+    const decodedToken= jwt.verify(request.token, config.SECRET)
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'invalid token'})
+    }
+
     await Blog.findByIdAndRemove(request.params.id)
         response.status(204).end()
   })
